feat: correctly process the edges when blurring
before, images being blurred would have the edges ignored and not processed, on multiple iterations it would become darker as the edges would be black.

Edges are now correctly included and the correct divisor is applied based on the number of included kernel values. e.g on edges not all NxN are included.